#2297 move helper function to separate file
